Create MA and SES.py
six-month moving average and simple exponential smoothing to obtain one-step-ahead forecasts of this series for the period 2015 to 2022. (Use the arithmetic average of the observations from 2000 to 2014 to initialize the exponential smoothing.) Your interface should allow the user to change input parameter α. In both cases (i.e., MA and ES forecasts), convert your forecasts of the percentage increase for the following year to a forecast of the GNP itself and compute the MAD and the MSE of the resulting forecasts.